requirements.txt
django>=4.0.3
pytz
requests
feedparser
beautifulsoup4
numpy
pandas
lxml
html5lib
django-extensions
djangorestframework
djoser
django-filter
openai>=1.0.0
ratelimit
scrapetube
django-import-export
langid
python-dotenv
supervisor
celery
flower
celery[redis]
django_celery_beat
django-webpush
django-pwa
sentry-sdk[django]
sentry-sdk[celery]
psycopg2-binary
